Download tile logs from S3 to workdir before reading
To produce the GeoTIFF output, we need to merge pre-processed log data
of the past 52 weeks. When we build the weekly logs, we store the
lines within in each file in a sorted order, so the Painter class can
do an efficient merge of the 52 weeks by scanning through the inputs
in parallel. However, some S3 object stores don't like it when a
client opens 52 connections and scans through their content in
parallel.

Apparently, when confronted with clients like ours, Ceph-based S3
object storage servers believe that they're under attack. As a
mitigation measure, the S3 server randomly closes connections. In our
client code, this connection shutdown manifests as random end-of-file
errors.

To work around this, we first download the 52 tile logs to our working
directory before reading them. By consequence, we need an additional
10G of empty local disk for the working directory to run the build
pipeline in production. Which is a lot better than crashing due to
random errors.